Artificial Intelligence (AI) Powered Virtual Personal Health Assistant Application to Help Overcome Mental Issues. (will continue to make better over time)
I basically wanted to create an Artificial intelligence powered virtual assistant application for people struggling with mental health issues. The app is supposed to act as a virtual companion, that basically gives them a therapist at their fingertips, the functionalities I was able to implement such as the chatbot companion to give them 24 hour support and someone to talk at any time they need it, the diaries functionality to allow them to write their thoughts in journal like form all diaries stored for them in a database and the map functionality this was to help patients find near by healthcare services if needed.